server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain
langchain-ollama
langchain-chroma
pandas
motor~=3.3.2
pymongo~=4.6.0
semantic-kernel[Ollama]
elevenlabs
python-dotenv
asyncio